Moved Connectivity check into BaseViewModel
